Task_1/requirements.txt
albumentations==0.5.2
click==8.1.3
matplotlib==3.4.1
nibabel==4.0.1
numpy==1.22.4
opencv_python_headless==4.5.1.48
pandas==1.1.4
Pillow==9.2.0
pyra_pytorch==1.3.0
scikit_learn==1.1.2
segmentation_models_pytorch==0.1.3
SimpleITK==2.2.0
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.60.0
wandb==0.13.2
